Add render optimizations and memory management improvements
- Enhanced `AvaloniaBitmapTwoLayerRenderTarget` and `SkiaCanvasTwoLayerRenderTarget` for improved pixel processing efficiency.
- Updated `Vic2Rasterizer` to utilize `uint` arrays for direct pixel manipulation, reducing casting overhead.
- Implemented zero-copy memory management with `NonOwningMemoryOwner` to optimize buffer handling in rendering.

+++ src/libraries/Highbyte.DotNet6502.Systems/Rendering/VideoFrameProvider/IVideoFrameProvider.cs
@@ -9,7 +9,7 @@ public interface IVideoFrameProvider : IRenderSource
     public event EventHandler? FrameCompleted;
 
     // Non-owning view to the *current* frame buffer (double-buffered internally)
-    public ReadOnlyMemory<byte> CurrentFrontBuffer { get; }
+    public ReadOnlyMemory<uint> CurrentFrontBuffer { get; }
     public int StrideBytes { get; }
 
     // Called by emulator to swap buffers at end-of-frame
